changed userInfo type to NSDictionary*
git-svn-id: svn://orca.physics.unc.edu/Orca@11827 fd86f90f-2c0a-422a-b972-84106d8db31b

+++ Source/Objects/Hardware/GPIB/Pulser33220/ORPulser33220Model.m
@@ -603,7 +603,7 @@ NSString* ORPulser33220ModelUSBInterfaceChanged = @"ORPulser33220ModelUSBInterfa
     }
 }
 
-- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(unsigned)inAmount
+- (void) netsocket:(NetSocket*)inNetSocket dataAvailable:(NSUInteger)inAmount
 {
     if(inNetSocket == socket){
 		NSString* theString = [[inNetSocket readString:NSASCIIStringEncoding] stringByTrimmingCharactersInSet:[NSCharacterSet whitespaceAndNewlineCharacterSet]];
